requirements.txt
django
pyrebase4